backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=42.0.0
python-multipart>=0.0.9
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
slowapi>=0.1.9
httpx>=0.27.0
Pillow>=10.0.0
pywebpush>=2.0.0
pytest>=7.0
pytest-asyncio>=0.21
